Redesign authenticated area with show routes and card grid index pages
Split worksheet/summary index pages into index (card grid with pagination)
and show (full content display) routes. Remove sidebar history sections and
per-request queries from HandleInertiaRequests. Unify dashboard recent
activity into a single merged feed.
