+++ resources/js/components/app-shell.tsx
@@ -1,6 +1,4 @@
 import { SidebarProvider } from '@/components/ui/sidebar';
-import { SharedData } from '@/types';
-import { usePage } from '@inertiajs/react';
 
 interface AppShellProps {
     children: React.ReactNode;
@@ -8,13 +6,11 @@ interface AppShellProps {
 }
 
 export function AppShell({ children, variant = 'header' }: AppShellProps) {
-    const isOpen = usePage<SharedData>().props.sidebarOpen;
-
     if (variant === 'header') {
         return (
             <div className="flex min-h-screen w-full flex-col">{children}</div>
         );
     }
 
-    return <SidebarProvider defaultOpen={isOpen}>{children}</SidebarProvider>;
+    return <SidebarProvider defaultOpen={true}>{children}</SidebarProvider>;
 }
