+++ resources/js/types/index.d.ts
@@ -26,20 +26,6 @@ export interface SharedData {
     name: string;
     quote: { message: string; author: string };
     auth: Auth;
-    sidebarOpen: boolean;
-    worksheetHistory?: {
-        id: number;
-        discipline: string;
-        topic: string;
-        created_at: string;
-    }[];
-    summaryHistory?: {
-        id: number;
-        title: string;
-        discipline: string;
-        topic: string;
-        created_at: string;
-    }[];
     [key: string]: unknown;
 }
